(refactor) Organize and modify files to serve vanilla create-react-app landing page
