+++ client/src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
-import './index.css';
+import App from './components/presentational/App';
+import './styles/css/index.css';
 
 ReactDOM.render(
   <App />,
-  document.getElementsByClassName('root')[0]
+  document.querySelector('.root')
 );
